+++ GapBuffer/iterator.h
@@ -5,6 +5,7 @@
 #include "Exception.h"
 #include "iterator_utilities.h"
 #include <vector>
+#include <stdint.h>
 
 /**
 * @brief iterator implementation for accessing characters, it skips a gap space like it doesn't exist.
@@ -56,11 +57,11 @@ class GapBuffer::iterator {
 
 	///Function tells us may our action of moving iterator cause out of range exception throwing.
 	template<typename GAPIt, typename It, typename, typename>
-	friend bool IsIterOutOfRange<GAPIt, It>(GAPIt, It, const char&, const int&);
+	friend bool IsIterOutOfRange<GAPIt, It>(GAPIt, It, const char&, const int64_t&);
 	
 	///Function tells us will we skip the gap if move our iterator.
 	template<typename GAPIt, typename>
-	friend bool WillSkipGap<GAPIt>(GAPIt, const char&, const int&);
+	friend bool WillSkipGap<GAPIt>(GAPIt, const char&, const int64_t&);
 
   private:
 	///Function tells us is our GapBuffer::iterator got into the gap space.
